Run bootscript from uboot config partition
The uboot boot.scr script should not be run from the root partition
as it is subject to breakage on update. Instead uboot to load it from
the environment fat partition.
We furthermore add the rauc configuration to the bootargs to be able to
determine if we booted from the bootscript.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>