imx: imx8m / imx7: Correctly identify bootdevice if fat env is in mmc
The fat env code when setup with CONFIG_ENV_FAT_INTERFACE="mmc" allows
for the CONFIG_ENV_FAT_DEVICE_AND_PART to be setup without a device when
set to e.g. ":2". It is then supposed to then check the bootdevice for
partition 2.
The i.MX platform code has a multitude of implementations of mmc_get_env_dev
to do this for the CONFIG_ENV_IS_IN_MMC case. If the env is in a fat which in
itself is in mmc this code will never be compiled in leading to the ENV device
falling back to CONFIG_SYS_MMC_ENV_DEV which is not boot device dependant.
To aid this we are adding the correct identifier functions to the CONFIG_MMC
compilation objects if the FAT interface is set to "mmc".
While this logic probably also applies to other i.MX platforms we are not changing
it here due to lack of validation oportunities.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>